renamed jcache sample configure-cache-types to runtime-cache-types
