+++ jcache/runtime-cache-types/src/main/java/org/javaee8/jcache/runtime/KeyValueService.java
@@ -1,4 +1,4 @@
-package org.javaee8.jcache.configuration;
+package org.javaee8.jcache.runtime;
 
 import javax.annotation.PostConstruct;
 import javax.annotation.PreDestroy;
