+++ jcache/runtime-cache-types/src/test/resources/beans.xml
