+++ jcache/runtime-cache-types/src/test/resources/jboss-deployment-structure.xml
